Canned responses for bug reports
This shows a 'reply' icon in the markup form header. The config to add more canned responses is
`maniphest.canned-responses`.

Note: There is no validation (other than json), so the config is expected to match the
required fields ('name' and 'message')

Note: This shows at all markups, so technically it can be used for patch
review and other places where canned reponses would fit.